caddyhttp: Implement `skip_log` handler
